feat: Enhance dialog utility with customizable width and scrollable content
feat: Register file_selector plugin for Linux

feat: Update CMake to include file_selector_windows plugin

feat: Register file_selector plugin for macOS

chore: Update pubspec.lock with new file_selector dependencies

feat: Add image_picker dependency for handling image selection

feat: Register file_selector plugin for Windows

feat: Implement data storage for halter calibration logs

feat: Implement data storage for halter device calibration

feat: Implement data storage for halter device calibration offsets

feat: Create model for halter calibration log

feat: Create model for halter device calibration

feat: Create model for halter device calibration offsets

feat: Implement controller for halter calibration logs

feat: Create UI page for displaying halter calibration logs

feat: Create detailed view for horse information with image carousel

+++ lib/app/modules/smart_halter/rule_engine/table/halter_table_rule_engine_controller.dart
@@ -236,9 +236,9 @@ class HalterTableRuleEngineController extends GetxController {
   }
 
   String positionClassify({
-    required int? pitch,
-    required int? roll,
-    required int? yaw,
+    required double? pitch,
+    required double? roll,
+    required double? yaw,
   }) {
     HalterPositionRuleEngineModel? bestMatch;
     int bestScore = -1;
